+++ src/services/apiPinsService.js
@@ -1,12 +1,14 @@
 import axios from 'axios';
 import { useSettingsStore } from '@/store/settingsStore';
+import { apiStore } from '@/store/store';
 
 const getBaseUrl = () => {
   const settingsStore = useSettingsStore();
+  const store = apiStore();
   const protocol = settingsStore.backendProtocol || 'http';
   const host = settingsStore.connection.ip || window.location.hostname;
   let port = settingsStore.connection.port || window.location.port || 80;
-  const apiPort = settingsStore.apiPort;
+  const apiPort = store.apiPort;
 
   // devport auf 5000 umleiten
   const isDev = process.env.NODE_ENV === 'development';
@@ -39,6 +41,12 @@ export default {
   getINDIDeviceList(device) {
     const { API_URL } = getUrls();
     return this._simpleGetRequest(`${API_URL}indi/${device}`);
+  },
+
+  //-------------------Focuser------------------------
+  focuserAction(action) {
+    const { BASE_URL } = getUrls();
+    return this._simpleGetRequest(`${BASE_URL}/equipment/focuser/${action}`);
   },
 
   //-------------------PHD2------------------------
@@ -141,6 +149,11 @@ export default {
   createPHD2Profile(name) {
     const { API_URL } = getUrls();
     return this._simplePostRequest(`${API_URL}phd2/profile/create`, { name });
+  },
+
+  getGuideCam() {
+    const { API_URL } = getUrls();
+    return this._simpleGetRequest(`${API_URL}phd2/camera/ids`);
   },
 
   // Private method for simple GET requests