+++ CHANGELOG.md
@@ -8,6 +8,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Navigation: Scroll indicators with fade effect and arrow hint when more icons are available
 - Camera: Separate Readout Mode settings for Image and Snap (API >= 2.2.14.3) - thanks to Chris
+- Camera: Add USB limit setting component for camera (API >= 2.2.14.3) - thanks to Chris
+- Focuser: Add use filter offsets setting
+- Plugin: ObservationPlaner - thanks to Flashy_DE (Alex)
 
 ### Changed
 - Camera: Unified UI design for settings components